% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmalob}
\alias{cfmalob}
\title{\code{cfmalob}}
\usage{
cfmalob(status, dbkey, objnam, class, freq, type, basis, observ, numobs,
  numchr, growth)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{class}{Integer}

\item{freq}{Integer}

\item{type}{Integer}

\item{basis}{Integer}

\item{observ}{Integer}

\item{numobs}{Integer}

\item{numchr}{Integer}

\item{growth}{Numeric}
}
\description{
Add a new object and specify its allocation.
}
