% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmasrt}
\alias{cfmasrt}
\title{\code{cfmasrt}}
\usage{
cfmasrt(status, connkey, assert_type, assertion, perspective, grouping,
  dblist)
}
\arguments{
\item{status}{Integer}

\item{connkey}{Integer}

\item{assert_type}{Integer}

\item{assertion}{Character}

\item{perspective}{Integer}

\item{grouping}{Integer}

\item{dblist}{Integer}
}
\description{
Check for database changes during a unit of work.
}
