% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmbwdy}
\alias{cfmbwdy}
\title{\code{cfmbwdy}}
\usage{
cfmbwdy(status, freq, date, biwkdy)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{biwkdy}{Integer}
}
\description{
Get the biweekday of a date.
}
