% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmchfr}
\alias{cfmchfr}
\title{\code{cfmchfr}}
\usage{
cfmchfr(status, sfreq, sdate, select, tfreq, tdate, relate)
}
\arguments{
\item{status}{Integer}

\item{sfreq}{Integer}

\item{sdate}{Integer}

\item{select}{Integer}

\item{tfreq}{Integer}

\item{tdate}{Integer}

\item{relate}{Integer}
}
\description{
Change a date to a different frequency.
}
