% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmcpob}
\alias{cfmcpob}
\title{\code{cfmcpob}}
\usage{
cfmcpob(status, srckey, tarkey, srcnam, tarnam)
}
\arguments{
\item{status}{Integer}

\item{srckey}{Integer}

\item{tarkey}{Integer}

\item{srcnam}{Character}

\item{tarnam}{Character}
}
\description{
Copy one object to another.
}
