% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdatd}
\alias{cfmdatd}
\title{\code{cfmdatd}}
\usage{
cfmdatd(status, freq, date, year, month, day)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{month}{Integer}

\item{day}{Integer}
}
\description{
Convert a date to a year, month, and day.
}
