% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdatf}
\alias{cfmdatf}
\title{\code{cfmdatf}}
\usage{
cfmdatf(status, freq, date, year, period, fmonth, flabel)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{period}{Integer}

\item{fmonth}{Integer}

\item{flabel}{Integer}
}
\description{
Convert a date to a fiscal year and period.
}
