% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdati}
\alias{cfmdati}
\title{\code{cfmdati}}
\usage{
cfmdati(status, freq, date, datstr, image, fmonth, flabel)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{datstr}{Character}

\item{image}{Character}

\item{fmonth}{Integer}

\item{flabel}{Integer}
}
\description{
Convert a date to a string using an image.
}
