% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdatl}
\alias{cfmdatl}
\title{\code{cfmdatl}}
\usage{
cfmdatl(status, freq, date, datstr, fmonth, flabel)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{datstr}{Character}

\item{fmonth}{Integer}

\item{flabel}{Integer}
}
\description{
Convert a date to a date literal string.
}
