% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdatp}
\alias{cfmdatp}
\title{\code{cfmdatp}}
\usage{
cfmdatp(status, freq, date, year, period)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{period}{Integer}
}
\description{
Convert a date to a year and period.
}
