% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdatt}
\alias{cfmdatt}
\title{\code{cfmdatt}}
\usage{
cfmdatt(status, freq, date, hour, minute, second, ddate)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{hour}{Integer}

\item{minute}{Integer}

\item{second}{Integer}

\item{ddate}{Integer}
}
\description{
Convert an intraday date to a daily date and time.
}
