% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmddat}
\alias{cfmddat}
\title{\code{cfmddat}}
\usage{
cfmddat(status, freq, date, year, month, day)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{month}{Integer}

\item{day}{Integer}
}
\description{
Convert a year, month, and day to a date.
}
