% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmddes}
\alias{cfmddes}
\title{\code{cfmddes}}
\usage{
cfmddes(status, dbkey, des)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{des}{Character}
}
\description{
Set a database's DESCRIPTION attribute.
}
