% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmddoc}
\alias{cfmddoc}
\title{\code{cfmddoc}}
\usage{
cfmddoc(status, dbkey, doc)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{doc}{Character}
}
\description{
Set a database's DOCUMENTATION attribute.
}
