% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdlen}
\alias{cfmdlen}
\title{\code{cfmdlen}}
\usage{
cfmdlen(status, dbkey, objnam, deslen, doclen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{deslen}{Integer}

\item{doclen}{Integer}
}
\description{
Get the length of an object's DESCRIPTION and DOCUMENTATION attributes.
}
