% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmdlob}
\alias{cfmdlob}
\title{\code{cfmdlob}}
\usage{
cfmdlob(status, dbkey, objnam)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}
}
\description{
Delete an object.
}
