% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmexpiration}
\alias{cfmexpiration}
\title{\code{cfmexpiration}}
\usage{
cfmexpiration(status, date)
}
\arguments{
\item{status}{Integer}

\item{date}{Integer}
}
\description{
Get the license expiration date.
}
