% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmfdiv}
\alias{cfmfdiv}
\title{\code{cfmfdiv}}
\usage{
cfmfdiv(status, freq1, freq2, flag)
}
\arguments{
\item{status}{Integer}

\item{freq1}{Integer}

\item{freq2}{Integer}

\item{flag}{Integer}
}
\description{
See if one frequency is evenly divisible by another.
}
