% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgcid}
\alias{cfmgcid}
\title{\code{cfmgcid}}
\usage{
cfmgcid(status, dbkey, connkey)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{connkey}{Integer}
}
\description{
Get the connection key for a database key.
}
