% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgdat}
\alias{cfmgdat}
\title{\code{cfmgdat}}
\usage{
cfmgdat(status, dbkey, objnam, freq, cdate, mdate)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{freq}{Integer}

\item{cdate}{Integer}

\item{mdate}{Integer}
}
\description{
Get an object's creation and last modification dates.
}
