% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgdba}
\alias{cfmgdba}
\title{\code{cfmgdba}}
\usage{
cfmgdba(status, dbkey, cyear, cmonth, cday, myear, mmonth, mday, desc, doc)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{cyear}{Integer}

\item{cmonth}{Integer}

\item{cday}{Integer}

\item{myear}{Integer}

\item{mmonth}{Integer}

\item{mday}{Integer}

\item{desc}{Character}

\item{doc}{Character}
}
\description{
Get database attributes.
}
