% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgdbd}
\alias{cfmgdbd}
\title{\code{cfmgdbd}}
\usage{
cfmgdbd(status, dbkey, freq, cdate, mdate)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{freq}{Integer}

\item{cdate}{Integer}

\item{mdate}{Integer}
}
\description{
Get a database's creation and last modification dates.
}
