% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmget_dimension}
\alias{cfmget_dimension}
\title{\code{cfmget_dimension}}
\usage{
cfmget_dimension(status, dbkey, dimen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{dimen}{Integer}
}
\description{
Get the dimension attribute for a database.
}
