% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmget_extradots}
\alias{cfmget_extradots}
\title{\code{cfmget_extradots}}
\usage{
cfmget_extradots(status, dbkey, xdots)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{xdots}{Integer}
}
\description{
Get the extradots attribute for a database.
}
