% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmglen}
\alias{cfmglen}
\title{\code{cfmglen}}
\usage{
cfmglen(status, dbkey, deslen, doclen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{deslen}{Integer}

\item{doclen}{Integer}
}
\description{
Get the length of a database's DESCRIPTION and DOCUMENTATION attributes.
}
