% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgnam}
\alias{cfmgnam}
\title{\code{cfmgnam}}
\usage{
cfmgnam(status, dbkey, objnam, basnam)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{basnam}{Character}
}
\description{
Get an object's real name.
}
