% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgsln}
\alias{cfmgsln}
\title{\code{cfmgsln}}
\usage{
cfmgsln(status, dbkey, objnam, length)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{length}{Integer}
}
\description{
Get an object's STRINGLENGTH attribute.
}
