% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgtali}
\alias{cfmgtali}
\title{\code{cfmgtali}}
\usage{
cfmgtali(status, dbkey, objnam, aliass, inlen, outlen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{aliass}{Character}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
Get an object's ALIAS attribute.
}
