% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgtaso}
\alias{cfmgtaso}
\title{\code{cfmgtaso}}
\usage{
cfmgtaso(status, dbkey, objnam, assoc, inlen, outlen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{assoc}{Character}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
Get an object's ASSOCIATE attribute.
}
