% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgtatt_double}
\alias{cfmgtatt_double}
\title{\code{cfmgtatt_double}}
\usage{
cfmgtatt_double(status, dbkey, objnam, atttyp, attnam, value, inlen,
  outlen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{atttyp}{Integer}

\item{attnam}{Character}

\item{value}{Numeric}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
Get the value of a user-defined \code{Numeric} attribute.
}
