% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgtatt_int}
\alias{cfmgtatt_int}
\title{\code{cfmgtatt_int}}
\usage{
cfmgtatt_int(status, dbkey, objnam, atttyp, attnam, value, inlen, outlen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{atttyp}{Integer}

\item{attnam}{Character}

\item{value}{Integer}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
Get the value of a user-defined \code{Integer} attribute.
}
