% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgtnl}
\alias{cfmgtnl}
\title{\code{cfmgtnl}}
\usage{
cfmgtnl(status, dbkey, objnam, index, strval, inlen, outlen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{index}{Integer}

\item{strval}{Character}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
Get an element of a namelist or the whole namelist.
}
