% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgtstr}
\alias{cfmgtstr}
\title{\code{cfmgtstr}}
\usage{
cfmgtstr(status, dbkey, objnam, range, strval, ismiss, inlen, outlen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{strval}{Character}

\item{ismiss}{Integer}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
Get a single string value.
}
