% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmgtsts}
\alias{cfmgtsts}
\title{\code{cfmgtsts}}
\usage{
cfmgtsts(status, dbkey, objnam, range, strary, misary, inlenary, outlenary)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{strary}{Character vector}

\item{misary}{Integer vector}

\item{inlenary}{Integer vector}

\item{outlenary}{Integer vector}
}
\description{
Get a range of string values.
}
