% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmidat}
\alias{cfmidat}
\title{\code{cfmidat}}
\usage{
cfmidat(status, freq, date, datstr, image, fmonth, flabel, centry)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{datstr}{Character}

\item{image}{Character}

\item{fmonth}{Integer}

\item{flabel}{Integer}

\item{centry}{Integer}
}
\description{
Convert a string to a date using an image.
}
