% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfminwc}
\alias{cfminwc}
\title{\code{cfminwc}}
\usage{
cfminwc(status, dbkey, wilnam)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{wilnam}{Character}
}
\description{
Initialize a wildcard expansion.
}
