% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmisbm}
\alias{cfmisbm}
\title{\code{cfmisbm}}
\usage{
cfmisbm(status, value, ismiss)
}
\arguments{
\item{status}{Integer}

\item{value}{Integer}

\item{ismiss}{Integer}
}
\description{
See if a Boolean value is a missing value.
}
