% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmisdm}
\alias{cfmisdm}
\title{\code{cfmisdm}}
\usage{
cfmisdm(status, value, ismiss)
}
\arguments{
\item{status}{Integer}

\item{value}{Integer}

\item{ismiss}{Integer}
}
\description{
See if a date value is a missing value.
}
