% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmislp}
\alias{cfmislp}
\title{\code{cfmislp}}
\usage{
cfmislp(status, year, leap)
}
\arguments{
\item{status}{Integer}

\item{year}{Integer}

\item{leap}{Integer}
}
\description{
See if a year is a leap year.
}
