% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmisnm}
\alias{cfmisnm}
\title{\code{cfmisnm}}
\usage{
cfmisnm(status, value, ismiss)
}
\arguments{
\item{status}{Integer}

\item{value}{Numeric}

\item{ismiss}{Integer}
}
\description{
See if a numeric value is a missing value.
}
