% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmispm}
\alias{cfmispm}
\title{\code{cfmispm}}
\usage{
cfmispm(status, value, ismiss)
}
\arguments{
\item{status}{Integer}

\item{value}{Numeric}

\item{ismiss}{Integer}
}
\description{
See if a precision value is a missing value.
}
