% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmissm}
\alias{cfmissm}
\title{\code{cfmissm}}
\usage{
cfmissm(status, value, ismiss)
}
\arguments{
\item{status}{Integer}

\item{value}{Character}

\item{ismiss}{Integer}
}
\description{
See if a string value is a missing value.
}
