% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmlali}
\alias{cfmlali}
\title{\code{cfmlali}}
\usage{
cfmlali(status, dbkey, objnam, alilen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{alilen}{Integer}
}
\description{
Get the length of an object's ALIAS attribute.
}
