% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmlaso}
\alias{cfmlaso}
\title{\code{cfmlaso}}
\usage{
cfmlaso(status, dbkey, objnam, asolen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{asolen}{Integer}
}
\description{
Get the length of an object's ASSOCIATE attribute.
}
