% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmlatt}
\alias{cfmlatt}
\title{\code{cfmlatt}}
\usage{
cfmlatt(status, dbkey, objnam, atttyp, attnam, attlen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{atttyp}{Integer}

\item{attnam}{Character}

\item{attlen}{Integer}
}
\description{
Get the length of a user-defined string or namelist attribute.
}
