% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmldat}
\alias{cfmldat}
\title{\code{cfmldat}}
\usage{
cfmldat(status, freq, date, datstr, fmonth, flabel, centry)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{datstr}{Character}

\item{fmonth}{Integer}

\item{flabel}{Integer}

\item{centry}{Integer}
}
\description{
Convert a date literal string to a date.
}
