% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmlsts}
\alias{cfmlsts}
\title{\code{cfmlsts}}
\usage{
cfmlsts(status, dbkey, objnam, range, lenary)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{lenary}{Integer vector}
}
\description{
Get the lengths of a range of string values.
}
