% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmncnt}
\alias{cfmncnt}
\title{\code{cfmncnt}}
\usage{
cfmncnt(status, dbkey, objnam, length)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{length}{Integer}
}
\description{
Get the length of a namelist.
}
