% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmnlen}
\alias{cfmnlen}
\title{\code{cfmnlen}}
\usage{
cfmnlen(status, dbkey, objnam, index, length)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{index}{Integer}

\item{length}{Integer}
}
\description{
Get the length of the buffer for reading a namelist.
}
