% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmnwob}
\alias{cfmnwob}
\title{\code{cfmnwob}}
\usage{
cfmnwob(status, dbkey, objnam, class, freq, type, basis, observ)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{class}{Integer}

\item{freq}{Integer}

\item{type}{Integer}

\item{basis}{Integer}

\item{observ}{Integer}
}
\description{
Add a new object to a database.
}
