% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmnxwc}
\alias{cfmnxwc}
\title{\code{cfmnxwc}}
\usage{
cfmnxwc(status, dbkey, objnam, class, type, freq)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{class}{Integer}

\item{type}{Integer}

\item{freq}{Integer}
}
\description{
Get the next name matching a wildcard.
}
