% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmopcn}
\alias{cfmopcn}
\title{\code{cfmopcn}}
\usage{
cfmopcn(status, connkey, service, hostname, username, password)
}
\arguments{
\item{status}{Integer}

\item{connkey}{Integer}

\item{service}{Character}

\item{hostname}{Character}

\item{username}{Character}

\item{password}{Character}
}
\description{
Open a connection to a FAME Database Server.
}
