% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmopdb}
\alias{cfmopdb}
\title{\code{cfmopdb}}
\usage{
cfmopdb(status, dbkey, dbname, mode)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{dbname}{Character}

\item{mode}{Integer}
}
\description{
Open a database.
}
