% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmopdc}
\alias{cfmopdc}
\title{\code{cfmopdc}}
\usage{
cfmopdc(status, dbkey, dbname, mode, connkey)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{dbname}{Character}

\item{mode}{Integer}

\item{connkey}{Integer}
}
\description{
Open a database on a connection.
}
