% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmoprc}
\alias{cfmoprc}
\title{\code{cfmoprc}}
\usage{
cfmoprc(status, dbkey, connkey)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{connkey}{Integer}
}
\description{
Open an analytical channel on a connection.
}
