% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmopre}
\alias{cfmopre}
\title{\code{cfmopre}}
\usage{
cfmopre(status, dbkey, svname)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{svname}{Character}
}
\description{
Open an analytical channel to a server.
}
