% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmosiz}
\alias{cfmosiz}
\title{\code{cfmosiz}}
\usage{
cfmosiz(status, dbkey, objnam, class, type, freq, fyear, fprd, lyear, lprd)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{class}{Integer}

\item{type}{Integer}

\item{freq}{Integer}

\item{fyear}{Integer}

\item{fprd}{Integer}

\item{lyear}{Integer}

\item{lprd}{Integer}
}
\description{
Get the size of an object.
}
