% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmpdat}
\alias{cfmpdat}
\title{\code{cfmpdat}}
\usage{
cfmpdat(status, freq, date, year, period)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{period}{Integer}
}
\description{
Convert a year and period to a date.
}
