% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmpind}
\alias{cfmpind}
\title{\code{cfmpind}}
\usage{
cfmpind(status, freq, count)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{count}{Integer}
}
\description{
Get the number of periods in a day.
}
