% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmpinm}
\alias{cfmpinm}
\title{\code{cfmpinm}}
\usage{
cfmpinm(status, freq, year, month, count)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{year}{Integer}

\item{month}{Integer}

\item{count}{Integer}
}
\description{
Get the number of periods in a month.
}
