% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmpiny}
\alias{cfmpiny}
\title{\code{cfmpiny}}
\usage{
cfmpiny(status, freq, year, count)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{year}{Integer}

\item{count}{Integer}
}
\description{
Get the number of periods in a year.
}
