% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmrdfa_float}
\alias{cfmrdfa_float}
\title{\code{cfmrdfa_float}}
\usage{
cfmrdfa_float(status, dbkey, objnam, wntobs, syear, sprd, gotobs, valary,
  tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{wntobs}{Integer}

\item{syear}{Integer}

\item{sprd}{Integer}

\item{gotobs}{Integer}

\item{valary}{Numeric vector}

\item{tmiss}{Integer}

\item{mistt}{Numeric vector}
}
\description{
Read all of the \code{Numeric} data from an object.
}
