% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmrdfa_int}
\alias{cfmrdfa_int}
\title{\code{cfmrdfa_int}}
\usage{
cfmrdfa_int(status, dbkey, objnam, wntobs, syear, sprd, gotobs, valary,
  tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{wntobs}{Integer}

\item{syear}{Integer}

\item{sprd}{Integer}

\item{gotobs}{Integer}

\item{valary}{Integer vector}

\item{tmiss}{Integer}

\item{mistt}{Integer vector}
}
\description{
Read all of the \code{Integer} data from an object.
}
