% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmrdfm}
\alias{cfmrdfm}
\title{\code{cfmrdfm}}
\usage{
cfmrdfm(status, dbkey, objnam, source, ilen, olen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{source}{Character}

\item{ilen}{Integer}

\item{olen}{Integer}
}
\description{
Read the source code for a formula from a database.
}
