% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmrmev}
\alias{cfmrmev}
\title{\code{cfmrmev}}
\usage{
cfmrmev(status, dbkey, expr, optns, wdbkey, objnam)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{expr}{Character}

\item{optns}{Character}

\item{wdbkey}{Integer}

\item{objnam}{Character}
}
\description{
Call on analytical channel to evaluate an expression.
}
