% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmrnob}
\alias{cfmrnob}
\title{\code{cfmrnob}}
\usage{
cfmrnob(status, dbkey, srcnam, tarnam)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{srcnam}{Character}

\item{tarnam}{Character}
}
\description{
Rename an object.
}
