% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmrrng_float}
\alias{cfmrrng_float}
\title{\code{cfmrrng_float}}
\usage{
cfmrrng_float(status, dbkey, objnam, range, valary, tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{valary}{Numeric vector}

\item{tmiss}{Integer}

\item{mistt}{Numeric vector}
}
\description{
Read a range of Numeric data from an object.
}
