% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmrrng_int}
\alias{cfmrrng_int}
\title{\code{cfmrrng_int}}
\usage{
cfmrrng_int(status, dbkey, objnam, range, valary, tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{valary}{Integer vector}

\item{tmiss}{Integer}

\item{mistt}{Integer vector}
}
\description{
Read a range of Integer data from an object.
}
