% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsali}
\alias{cfmsali}
\title{\code{cfmsali}}
\usage{
cfmsali(status, dbkey, objnam, aliass)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{aliass}{Character}
}
\description{
Set an object's ALIAS attribute.
}
