% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsaso}
\alias{cfmsaso}
\title{\code{cfmsaso}}
\usage{
cfmsaso(status, dbkey, objnam, assoc)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{assoc}{Character}
}
\description{
Set an object's ASSOCIATE attribute.
}
