% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsatt_char}
\alias{cfmsatt_char}
\title{\code{cfmsatt_char}}
\usage{
cfmsatt_char(status, dbkey, objnam, atttyp, attnam, value)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{atttyp}{Integer}

\item{attnam}{Character}

\item{value}{Character}
}
\description{
Set the value of a user-defined Character attribute
}
