% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsatt_double}
\alias{cfmsatt_double}
\title{\code{cfmsatt_double}}
\usage{
cfmsatt_double(status, dbkey, objnam, atttyp, attnam, value)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{atttyp}{Integer}

\item{attnam}{Character}

\item{value}{Numeric}
}
\description{
Set the value of a user-defined Numeric attribute
}
