% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsatt_int}
\alias{cfmsatt_int}
\title{\code{cfmsatt_int}}
\usage{
cfmsatt_int(status, dbkey, objnam, atttyp, attnam, value)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{atttyp}{Integer}

\item{attnam}{Character}

\item{value}{Integer}
}
\description{
Set the value of a user-defined Integer attribute
}
