% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsbas}
\alias{cfmsbas}
\title{\code{cfmsbas}}
\usage{
cfmsbas(status, dbkey, objnam, basis)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{basis}{Integer}
}
\description{
Set an object's BASIS attribute.
}
