% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsbm}
\alias{cfmsbm}
\title{\code{cfmsbm}}
\usage{
cfmsbm(status, nctran, ndtran, natran, bmistt)
}
\arguments{
\item{status}{Integer}

\item{nctran}{Integer}

\item{ndtran}{Integer}

\item{natran}{Integer}

\item{bmistt}{Integer vector}
}
\description{
Set a boolean missing value translation table.
}
