% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsdes}
\alias{cfmsdes}
\title{\code{cfmsdes}}
\usage{
cfmsdes(status, dbkey, objnam, desc)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{desc}{Character}
}
\description{
Set an object's DESCRIPTION attribute.
}
