% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsdm}
\alias{cfmsdm}
\title{\code{cfmsdm}}
\usage{
cfmsdm(status, nctran, ndtran, natran, dmistt)
}
\arguments{
\item{status}{Integer}

\item{nctran}{Integer}

\item{ndtran}{Integer}

\item{natran}{Integer}

\item{dmistt}{Integer vector}
}
\description{
Set a date missing value translation table.
}
