% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsdoc}
\alias{cfmsdoc}
\title{\code{cfmsdoc}}
\usage{
cfmsdoc(status, dbkey, objnam, doc)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{doc}{Character}
}
\description{
Set an object's DOCUMENTATION attribute.
}
