% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmset_dimension}
\alias{cfmset_dimension}
\title{\code{cfmset_dimension}}
\usage{
cfmset_dimension(status, dbkey, dimen)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{dimen}{Integer}
}
\description{
Set the dimension attribute for a database.
}
