% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmset_extradots}
\alias{cfmset_extradots}
\title{\code{cfmset_extradots}}
\usage{
cfmset_extradots(status, dbkey, xdots)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{xdots}{Integer}
}
\description{
Set the extradots attribute for a database.
}
