% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsfis}
\alias{cfmsfis}
\title{\code{cfmsfis}}
\usage{
cfmsfis(status, freq, syear, sprd, eyear, eprd, range, numobs, fmonth,
  flabel)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{syear}{Integer}

\item{sprd}{Integer}

\item{eyear}{Integer}

\item{eprd}{Integer}

\item{range}{Integer vector}

\item{numobs}{Integer}

\item{fmonth}{Integer}

\item{flabel}{Integer}
}
\description{
Set a fiscal range for reading or writing data.
}
