% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsnm}
\alias{cfmsnm}
\title{\code{cfmsnm}}
\usage{
cfmsnm(status, nctran, ndtran, natran, nmistt)
}
\arguments{
\item{status}{Integer}

\item{nctran}{Numeric}

\item{ndtran}{Numeric}

\item{natran}{Numeric}

\item{nmistt}{Numeric vector}
}
\description{
Set a numeric missing value translation table.
}
