% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsobs}
\alias{cfmsobs}
\title{\code{cfmsobs}}
\usage{
cfmsobs(status, dbkey, objnam, observ)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{observ}{Integer}
}
\description{
Set an object's OBSERVED attribute.
}
