% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsopt}
\alias{cfmsopt}
\title{\code{cfmsopt}}
\usage{
cfmsopt(status, optnam, optval)
}
\arguments{
\item{status}{Integer}

\item{optnam}{Character}

\item{optval}{Character}
}
\description{
Setting options in the C HLI.
}
