% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmspm}
\alias{cfmspm}
\title{\code{cfmspm}}
\usage{
cfmspm(status, nctran, ndtran, natran, pmistt)
}
\arguments{
\item{status}{Integer}

\item{nctran}{Numeric}

\item{ndtran}{Numeric}

\item{natran}{Numeric}

\item{pmistt}{Numeric vector}
}
\description{
Set a precision missing value translation table.
}
