% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmsrng}
\alias{cfmsrng}
\title{\code{cfmsrng}}
\usage{
cfmsrng(status, freq, syear, sprd, eyear, eprd, range, numobs)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{syear}{Integer}

\item{sprd}{Integer}

\item{eyear}{Integer}

\item{eprd}{Integer}

\item{range}{Integer vector}

\item{numobs}{Integer}
}
\description{
Set a range for reading or writing data.
}
