% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmssln}
\alias{cfmssln}
\title{\code{cfmssln}}
\usage{
cfmssln(status, dbkey, objnam, length)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{length}{Integer}
}
\description{
Set an object's STRINGLENGTH attribute.
}
