% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmtdat}
\alias{cfmtdat}
\title{\code{cfmtdat}}
\usage{
cfmtdat(status, freq, date, hour, minute, second, ddate)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{hour}{Integer}

\item{minute}{Integer}

\item{second}{Integer}

\item{ddate}{Integer}
}
\description{
Convert a daily date and time to an intraday date.
}
