% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmufrq}
\alias{cfmufrq}
\title{\code{cfmufrq}}
\usage{
cfmufrq(status, freq, base, nunits, year, month)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{base}{Integer}

\item{nunits}{Integer}

\item{year}{Integer}

\item{month}{Integer}
}
\description{
Unpack a frequency.
}
