% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmver}
\alias{cfmver}
\title{\code{cfmver}}
\usage{
cfmver(status, ver)
}
\arguments{
\item{status}{Integer}

\item{ver}{Numeric}
}
\description{
Get the C HLI version number.
}
\examples{
status <- Integer(-1L)
ver <- Numeric(-1)
cfmver(status,ver)
cat(sprintf("CHLI version \%.5f",ver$value))

}
