% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwhat}
\alias{cfmwhat}
\title{\code{cfmwhat}}
\usage{
cfmwhat(status, dbkey, objnam, class, type, freq, basis, observ, fyear,
  fprd, lyear, lprd, cyear, cmonth, cday, myear, mmonth, mday, desc, doc)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{class}{Integer}

\item{type}{Integer}

\item{freq}{Integer}

\item{basis}{Integer}

\item{observ}{Integer}

\item{fyear}{Integer}

\item{fprd}{Integer}

\item{lyear}{Integer}

\item{lprd}{Integer}

\item{cyear}{Integer}

\item{cmonth}{Integer}

\item{cday}{Integer}

\item{myear}{Integer}

\item{mmonth}{Integer}

\item{mday}{Integer}

\item{desc}{Character}

\item{doc}{Character}
}
\description{
Get information about an object.
}
