% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwkdy}
\alias{cfmwkdy}
\title{\code{cfmwkdy}}
\usage{
cfmwkdy(status, freq, date, wkdy)
}
\arguments{
\item{status}{Integer}

\item{freq}{Integer}

\item{date}{Integer}

\item{wkdy}{Integer}
}
\description{
Get the weekday of a date.
}
