% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwrmt_double}
\alias{cfmwrmt_double}
\title{\code{cfmwrmt_double}}
\usage{
cfmwrmt_double(status, dbkey, objnam, objtyp, range, valary, tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{objtyp}{Integer}

\item{range}{Integer vector}

\item{valary}{Numeric vector}

\item{tmiss}{Integer}

\item{mistt}{Numeric vector}
}
\description{
Write a range of Numeric data to a remote database.
}
