% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwrmt_int}
\alias{cfmwrmt_int}
\title{\code{cfmwrmt_int}}
\usage{
cfmwrmt_int(status, dbkey, objnam, objtyp, range, valary, tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{objtyp}{Integer}

\item{range}{Integer vector}

\item{valary}{Integer vector}

\item{tmiss}{Integer}

\item{mistt}{Integer vector}
}
\description{
Write a range of Integer data to a remote database.
}
