% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwrng_double}
\alias{cfmwrng_double}
\title{\code{cfmwrng_double}}
\usage{
cfmwrng_double(status, dbkey, objnam, range, valary, tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{valary}{Numeric vector}

\item{tmiss}{Integer}

\item{mistt}{Numeric vector}
}
\description{
Write a range of Numeric data.
}
