% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwrng_int}
\alias{cfmwrng_int}
\title{\code{cfmwrng_int}}
\usage{
cfmwrng_int(status, dbkey, objnam, range, valary, tmiss, mistt)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{valary}{Integer vector}

\item{tmiss}{Integer}

\item{mistt}{Integer vector}
}
\description{
Write a range of Integer data.
}
