% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwstr}
\alias{cfmwstr}
\title{\code{cfmwstr}}
\usage{
cfmwstr(status, dbkey, objnam, range, strval, ismiss, length)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{strval}{Character}

\item{ismiss}{Integer}

\item{length}{Integer}
}
\description{
Write a single string value.
}
