% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwsts}
\alias{cfmwsts}
\title{\code{cfmwsts}}
\usage{
cfmwsts(status, dbkey, objnam, range, strary, misary, lenary)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{strary}{Character vector}

\item{misary}{Integer vector}

\item{lenary}{Integer vector}
}
\description{
Write a range of string values.
}
