% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{cfmwtnl}
\alias{cfmwtnl}
\title{\code{cfmwtnl}}
\usage{
cfmwtnl(status, dbkey, objnam, index, strval)
}
\arguments{
\item{status}{Integer}

\item{dbkey}{Integer}

\item{objnam}{Character}

\item{index}{Integer}

\item{strval}{Character}
}
\description{
Write an element of a namelist or a whole namelist.
}
