% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{class}
\name{Character-class}
\alias{Character-class}
\alias{Character}
\title{\code{Character}}
\description{
A reference class to implement mutable character data.
}
\section{Fields}{

\describe{
\item{\code{value}}{A character vector.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(s = "")}}{Initialize a character vector.}
}}

\examples{
blankLine <- Character(sprintf("\%80s",""))

}
