% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_biweekday}
\alias{fame_biweekday}
\title{C-HLI function \code{fame_biweekday}}
\usage{
fame_biweekday(freq, date, biweekday)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{biweekday}{Integer}
}
\description{
C-HLI function \code{fame_biweekday}
}
