% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_current_time}
\alias{fame_current_time}
\title{C-HLI function \code{fame_current_time}}
\usage{
fame_current_time(freq, date)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}
}
\description{
C-HLI function \code{fame_current_time}
}
