% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_date_missing_type}
\alias{fame_date_missing_type}
\title{C-HLI function \code{fame_date_missing_type}}
\usage{
fame_date_missing_type(date, missing_type)
}
\arguments{
\item{date}{Integer}

\item{missing_type}{Integer}
}
\description{
C-HLI function \code{fame_date_missing_type}
}
