% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_date_to_literal}
\alias{fame_date_to_literal}
\title{C-HLI function \code{fame_date_to_literal}}
\usage{
fame_date_to_literal(freq, date, literal, end_month, label_year, inlen,
  outlen)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{literal}{Character}

\item{end_month}{Integer}

\item{label_year}{Integer}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
C-HLI function \code{fame_date_to_literal}
}
