% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_dateof}
\alias{fame_dateof}
\title{C-HLI function \code{fame_dateof}}
\usage{
fame_dateof(from_freq, from_date, select, to_freq, to_date, relate)
}
\arguments{
\item{from_freq}{Integer}

\item{from_date}{Integer}

\item{select}{Integer}

\item{to_freq}{Integer}

\item{to_date}{Integer}

\item{relate}{Integer}
}
\description{
C-HLI function \code{fame_dateof}
}
