% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_day_to_index}
\alias{fame_day_to_index}
\title{C-HLI function \code{fame_day_to_index}}
\usage{
fame_day_to_index(freq, date, year, month, day, relate)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{month}{Integer}

\item{day}{Integer}

\item{relate}{Integer}
}
\description{
C-HLI function \code{fame_day_to_index}
}
