% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_dimchar}
\alias{fame_dimchar}
\title{C-HLI function \code{fame_dimchar}}
\usage{
fame_dimchar(dimchar, dimnum)
}
\arguments{
\item{dimchar}{Character}

\item{dimnum}{Integer}
}
\description{
C-HLI function \code{fame_dimchar}
}
