% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_dimnum}
\alias{fame_dimnum}
\title{C-HLI function \code{fame_dimnum}}
\usage{
fame_dimnum(dimnum, dimchar)
}
\arguments{
\item{dimnum}{Integer}

\item{dimchar}{Character}
}
\description{
C-HLI function \code{fame_dimnum}
}
