% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_expiration_date}
\alias{fame_expiration_date}
\title{C-HLI function \code{fame_expiration_date}}
\usage{
fame_expiration_date(date)
}
\arguments{
\item{date}{Integer}
}
\description{
C-HLI function \code{fame_expiration_date}
}
