% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_format_date}
\alias{fame_format_date}
\title{C-HLI function \code{fame_format_date}}
\usage{
fame_format_date(freq, date, string, image, end_month, label_year, inlen,
  outlen)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{string}{Character}

\item{image}{Character}

\item{end_month}{Integer}

\item{label_year}{Integer}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
C-HLI function \code{fame_format_date}
}
