% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_free_wildcard}
\alias{fame_free_wildcard}
\title{C-HLI function \code{fame_free_wildcard}}
\usage{
fame_free_wildcard(wildkey)
}
\arguments{
\item{wildkey}{Integer}
}
\description{
C-HLI function \code{fame_free_wildcard}
}
