% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_base_frequency}
\alias{fame_get_base_frequency}
\title{C-HLI function \code{fame_get_base_frequency}}
\usage{
fame_get_base_frequency(base_freq, freq)
}
\arguments{
\item{base_freq}{Integer}

\item{freq}{Integer}
}
\description{
C-HLI function \code{fame_get_base_frequency}
}
