% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_db_dates}
\alias{fame_get_db_dates}
\title{C-HLI function \code{fame_get_db_dates}}
\usage{
fame_get_db_dates(dbkey, cdate, mdate)
}
\arguments{
\item{dbkey}{Integer}

\item{cdate}{Integer}

\item{mdate}{Integer}
}
\description{
C-HLI function \code{fame_get_db_dates}
}
