% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_dbversion}
\alias{fame_get_dbversion}
\title{C-HLI function \code{fame_get_dbversion}}
\usage{
fame_get_dbversion(dbkey, dbversion)
}
\arguments{
\item{dbkey}{Integer}

\item{dbversion}{Integer}
}
\description{
C-HLI function \code{fame_get_dbversion}
}
