% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_namelist_attribute}
\alias{fame_get_namelist_attribute}
\title{C-HLI function \code{fame_get_namelist_attribute}}
\usage{
fame_get_namelist_attribute(dbkey, objnam, attnam, value, inlen, outlen)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{attnam}{Character}

\item{value}{Character}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
C-HLI function \code{fame_get_namelist_attribute}
}
