% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_next_wildcard}
\alias{fame_get_next_wildcard}
\title{C-HLI function \code{fame_get_next_wildcard}}
\usage{
fame_get_next_wildcard(wildkey, objname, obj_class, type, freq, start, end,
  inlen, outlen)
}
\arguments{
\item{wildkey}{Integer}

\item{objname}{Character}

\item{obj_class}{Integer}

\item{type}{Integer}

\item{freq}{Integer}

\item{start}{Integer}

\item{end}{Integer}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
C-HLI function \code{fame_get_next_wildcard}
}
