% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_numerics}
\alias{fame_get_numerics}
\title{C-HLI function \code{fame_get_numerics}}
\usage{
fame_get_numerics(dbkey, objnam, range, valary)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{valary}{Numeric vector}
}
\description{
C-HLI function \code{fame_get_numerics}
}
