% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_range_numobs}
\alias{fame_get_range_numobs}
\title{C-HLI function \code{fame_get_range_numobs}}
\usage{
fame_get_range_numobs(range, numobs)
}
\arguments{
\item{range}{Integer vector}

\item{numobs}{Integer}
}
\description{
C-HLI function \code{fame_get_range_numobs}
}
