% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_real_name}
\alias{fame_get_real_name}
\title{C-HLI function \code{fame_get_real_name}}
\usage{
fame_get_real_name(dbkey, objname, basename, inlen, outlen)
}
\arguments{
\item{dbkey}{Integer}

\item{objname}{Character}

\item{basename}{Character}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
C-HLI function \code{fame_get_real_name}
}
