% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_server_input}
\alias{fame_get_server_input}
\title{C-HLI function \code{fame_get_server_input}}
\usage{
fame_get_server_input(buffer, buffer_size, input_size)
}
\arguments{
\item{buffer}{Character}

\item{buffer_size}{Integer}

\item{input_size}{Integer}
}
\description{
C-HLI function \code{fame_get_server_input}
}
