% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_get_strings}
\alias{fame_get_strings}
\title{C-HLI function \code{fame_get_strings}}
\usage{
fame_get_strings(dbkey, objnam, range, strary, inlenary, outlenary)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{strary}{Character vector}

\item{inlenary}{Integer vector}

\item{outlenary}{Integer vector}
}
\description{
C-HLI function \code{fame_get_strings}
}
