% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_index_to_day}
\alias{fame_index_to_day}
\title{C-HLI function \code{fame_index_to_day}}
\usage{
fame_index_to_day(freq, date, year, month, day)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{month}{Integer}

\item{day}{Integer}
}
\description{
C-HLI function \code{fame_index_to_day}
}
