% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_index_to_int}
\alias{fame_index_to_int}
\title{C-HLI function \code{fame_index_to_int}}
\usage{
fame_index_to_int(indexes64ary, indexes32ary, numobs)
}
\arguments{
\item{indexes64ary}{Integer vector}

\item{indexes32ary}{Integer vector}

\item{numobs}{Integer}
}
\description{
C-HLI function \code{fame_index_to_int}
}
