% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_index_to_time}
\alias{fame_index_to_time}
\title{C-HLI function \code{fame_index_to_time}}
\usage{
fame_index_to_time(freq, date, year, month, day, hour, minute, second,
  millisecond)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{month}{Integer}

\item{day}{Integer}

\item{hour}{Integer}

\item{minute}{Integer}

\item{second}{Integer}

\item{millisecond}{Integer}
}
\description{
C-HLI function \code{fame_index_to_time}
}
