% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_index_to_year_period}
\alias{fame_index_to_year_period}
\title{C-HLI function \code{fame_index_to_year_period}}
\usage{
fame_index_to_year_period(freq, date, year, period)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{period}{Integer}
}
\description{
C-HLI function \code{fame_index_to_year_period}
}
