% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_info}
\alias{fame_info}
\title{C-HLI function \code{fame_info}}
\usage{
fame_info(dbkey, oname, oclass, type, freq, findex, lindex, basis, observ,
  cdate, mdate, desc, indesclen, outdesclen, doc, indoclen, outdoclen)
}
\arguments{
\item{dbkey}{Integer}

\item{oname}{Character}

\item{oclass}{Integer}

\item{type}{Integer}

\item{freq}{Integer}

\item{findex}{Integer}

\item{lindex}{Integer}

\item{basis}{Integer}

\item{observ}{Integer}

\item{cdate}{Integer}

\item{mdate}{Integer}

\item{desc}{Character}

\item{indesclen}{Integer}

\item{outdesclen}{Integer}

\item{doc}{Character}

\item{indoclen}{Integer}

\item{outdoclen}{Integer}
}
\description{
C-HLI function \code{fame_info}
}
