% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_init_range_from_end_numobs}
\alias{fame_init_range_from_end_numobs}
\title{C-HLI function \code{fame_init_range_from_end_numobs}}
\usage{
fame_init_range_from_end_numobs(range, freq, end, numobs)
}
\arguments{
\item{range}{Integer vector}

\item{freq}{Integer}

\item{end}{Integer}

\item{numobs}{Integer}
}
\description{
C-HLI function \code{fame_init_range_from_end_numobs}
}
