% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_init_range_from_indexes}
\alias{fame_init_range_from_indexes}
\title{C-HLI function \code{fame_init_range_from_indexes}}
\usage{
fame_init_range_from_indexes(range, freq, start, end)
}
\arguments{
\item{range}{Integer vector}

\item{freq}{Integer}

\item{start}{Integer}

\item{end}{Integer}
}
\description{
C-HLI function \code{fame_init_range_from_indexes}
}
