% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_init_wildcard}
\alias{fame_init_wildcard}
\title{C-HLI function \code{fame_init_wildcard}}
\usage{
fame_init_wildcard(dbkey, wildkey, wildname, wildonly, wildstart)
}
\arguments{
\item{dbkey}{Integer}

\item{wildkey}{Integer}

\item{wildname}{Character}

\item{wildonly}{Integer}

\item{wildstart}{Character}
}
\description{
C-HLI function \code{fame_init_wildcard}
}
