% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_int_to_index}
\alias{fame_int_to_index}
\title{C-HLI function \code{fame_int_to_index}}
\usage{
fame_int_to_index(indexes32, indexes64, numobs)
}
\arguments{
\item{indexes32}{Integer}

\item{indexes64}{Integer}

\item{numobs}{Integer}
}
\description{
C-HLI function \code{fame_int_to_index}
}
