% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_len_strings}
\alias{fame_len_strings}
\title{C-HLI function \code{fame_len_strings}}
\usage{
fame_len_strings(dbkey, objnam, range, lenary)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{lenary}{Integer vector}
}
\description{
C-HLI function \code{fame_len_strings}
}
