% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_literal_to_date}
\alias{fame_literal_to_date}
\title{C-HLI function \code{fame_literal_to_date}}
\usage{
fame_literal_to_date(freq, date, literal, end_month, label_year, century)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{literal}{Character}

\item{end_month}{Integer}

\item{label_year}{Integer}

\item{century}{Integer}
}
\description{
C-HLI function \code{fame_literal_to_date}
}
