% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_modify_formula}
\alias{fame_modify_formula}
\title{C-HLI function \code{fame_modify_formula}}
\usage{
fame_modify_formula(dbkey, oname, value)
}
\arguments{
\item{dbkey}{Integer}

\item{oname}{Character}

\item{value}{Character}
}
\description{
C-HLI function \code{fame_modify_formula}
}
