% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_parse_frequency}
\alias{fame_parse_frequency}
\title{C-HLI function \code{fame_parse_frequency}}
\usage{
fame_parse_frequency(freq, frequency_spec)
}
\arguments{
\item{freq}{Integer}

\item{frequency_spec}{Character}
}
\description{
C-HLI function \code{fame_parse_frequency}
}
