% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_quick_info}
\alias{fame_quick_info}
\title{C-HLI function \code{fame_quick_info}}
\usage{
fame_quick_info(dbkey, oname, oclass, type, freq, findex, lindex)
}
\arguments{
\item{dbkey}{Integer}

\item{oname}{Character}

\item{oclass}{Integer}

\item{type}{Integer}

\item{freq}{Integer}

\item{findex}{Integer}

\item{lindex}{Integer}
}
\description{
C-HLI function \code{fame_quick_info}
}
