% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_remexec}
\alias{fame_remexec}
\title{C-HLI function \code{fame_remexec}}
\usage{
fame_remexec(dbkey, command, wdbkey, objnam)
}
\arguments{
\item{dbkey}{Integer}

\item{command}{Character}

\item{wdbkey}{Integer}

\item{objnam}{Character}
}
\description{
C-HLI function \code{fame_remexec}
}
