% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_scan_date}
\alias{fame_scan_date}
\title{C-HLI function \code{fame_scan_date}}
\usage{
fame_scan_date(freq, date, string, image, end_month, label_year, century)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{string}{Character}

\item{image}{Character}

\item{end_month}{Integer}

\item{label_year}{Integer}

\item{century}{Integer}
}
\description{
C-HLI function \code{fame_scan_date}
}
