% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_set_date_attribute}
\alias{fame_set_date_attribute}
\title{C-HLI function \code{fame_set_date_attribute}}
\usage{
fame_set_date_attribute(dbkey, objnam, attnam, value, freq)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{attnam}{Character}

\item{value}{Integer}

\item{freq}{Integer}
}
\description{
C-HLI function \code{fame_set_date_attribute}
}
