% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_set_precision_attribute}
\alias{fame_set_precision_attribute}
\title{C-HLI function \code{fame_set_precision_attribute}}
\usage{
fame_set_precision_attribute(dbkey, objnam, attnam, value)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{attnam}{Character}

\item{value}{Numeric}
}
\description{
C-HLI function \code{fame_set_precision_attribute}
}
