% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_time_to_index}
\alias{fame_time_to_index}
\title{C-HLI function \code{fame_time_to_index}}
\usage{
fame_time_to_index(freq, date, year, month, day, hour, minute, second,
  millisecond, relate)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{month}{Integer}

\item{day}{Integer}

\item{hour}{Integer}

\item{minute}{Integer}

\item{second}{Integer}

\item{millisecond}{Integer}

\item{relate}{Integer}
}
\description{
C-HLI function \code{fame_time_to_index}
}
