% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_type_to_string}
\alias{fame_type_to_string}
\title{C-HLI function \code{fame_type_to_string}}
\usage{
fame_type_to_string(type, type_spec, inlen, outlen)
}
\arguments{
\item{type}{Integer}

\item{type_spec}{Character}

\item{inlen}{Integer}

\item{outlen}{Integer}
}
\description{
C-HLI function \code{fame_type_to_string}
}
