% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_weekday}
\alias{fame_weekday}
\title{C-HLI function \code{fame_weekday}}
\usage{
fame_weekday(freq, date, weekday)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{weekday}{Integer}
}
\description{
C-HLI function \code{fame_weekday}
}
