% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_write_booleans}
\alias{fame_write_booleans}
\title{C-HLI function \code{fame_write_booleans}}
\usage{
fame_write_booleans(dbkey, objnam, range, valary)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{valary}{Integer vector}
}
\description{
C-HLI function \code{fame_write_booleans}
}
