% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_write_dates}
\alias{fame_write_dates}
\title{C-HLI function \code{fame_write_dates}}
\usage{
fame_write_dates(dbkey, objnam, range, value_type, valary)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{value_type}{Integer}

\item{valary}{Integer vector}
}
\description{
C-HLI function \code{fame_write_dates}
}
