% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_write_precisions}
\alias{fame_write_precisions}
\title{C-HLI function \code{fame_write_precisions}}
\usage{
fame_write_precisions(dbkey, objnam, range, valary)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{valary}{Numeric vector}
}
\description{
C-HLI function \code{fame_write_precisions}
}
