% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_write_strings}
\alias{fame_write_strings}
\title{C-HLI function \code{fame_write_strings}}
\usage{
fame_write_strings(dbkey, objnam, range, strary)
}
\arguments{
\item{dbkey}{Integer}

\item{objnam}{Character}

\item{range}{Integer vector}

\item{strary}{Character vector}
}
\description{
C-HLI function \code{fame_write_strings}
}
