% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\name{fame_year_period_to_index}
\alias{fame_year_period_to_index}
\title{C-HLI function \code{fame_year_period_to_index}}
\usage{
fame_year_period_to_index(freq, date, year, period)
}
\arguments{
\item{freq}{Integer}

\item{date}{Integer}

\item{year}{Integer}

\item{period}{Integer}
}
\description{
C-HLI function \code{fame_year_period_to_index}
}
