% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HABORT}
\alias{HABORT}
\title{C-HLI constant \code{HABORT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HABORT
}
\description{
C-HLI constant \code{HABORT}
}
\keyword{datasets}
