% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HAFTER}
\alias{HAFTER}
\title{C-HLI constant \code{HAFTER}}
\format{An object of class \code{integer} of length 1.}
\usage{
HAFTER
}
\description{
C-HLI constant \code{HAFTER}
}
\keyword{datasets}
