% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANAPR}
\alias{HANAPR}
\title{C-HLI constant \code{HANAPR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANAPR
}
\description{
C-HLI constant \code{HANAPR}
}
\keyword{datasets}
