% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANAUG}
\alias{HANAUG}
\title{C-HLI constant \code{HANAUG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANAUG
}
\description{
C-HLI constant \code{HANAUG}
}
\keyword{datasets}
