% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANDEC}
\alias{HANDEC}
\title{C-HLI constant \code{HANDEC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANDEC
}
\description{
C-HLI constant \code{HANDEC}
}
\keyword{datasets}
