% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANFEB}
\alias{HANFEB}
\title{C-HLI constant \code{HANFEB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANFEB
}
\description{
C-HLI constant \code{HANFEB}
}
\keyword{datasets}
