% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANJAN}
\alias{HANJAN}
\title{C-HLI constant \code{HANJAN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANJAN
}
\description{
C-HLI constant \code{HANJAN}
}
\keyword{datasets}
