% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANJUL}
\alias{HANJUL}
\title{C-HLI constant \code{HANJUL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANJUL
}
\description{
C-HLI constant \code{HANJUL}
}
\keyword{datasets}
