% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANJUN}
\alias{HANJUN}
\title{C-HLI constant \code{HANJUN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANJUN
}
\description{
C-HLI constant \code{HANJUN}
}
\keyword{datasets}
