% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANMAR}
\alias{HANMAR}
\title{C-HLI constant \code{HANMAR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANMAR
}
\description{
C-HLI constant \code{HANMAR}
}
\keyword{datasets}
