% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANMAY}
\alias{HANMAY}
\title{C-HLI constant \code{HANMAY}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANMAY
}
\description{
C-HLI constant \code{HANMAY}
}
\keyword{datasets}
