% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANNOV}
\alias{HANNOV}
\title{C-HLI constant \code{HANNOV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANNOV
}
\description{
C-HLI constant \code{HANNOV}
}
\keyword{datasets}
