% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANOCT}
\alias{HANOCT}
\title{C-HLI constant \code{HANOCT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANOCT
}
\description{
C-HLI constant \code{HANOCT}
}
\keyword{datasets}
