% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HANSEP}
\alias{HANSEP}
\title{C-HLI constant \code{HANSEP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HANSEP
}
\description{
C-HLI constant \code{HANSEP}
}
\keyword{datasets}
