% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HAPOST}
\alias{HAPOST}
\title{C-HLI constant \code{HAPOST}}
\format{An object of class \code{integer} of length 1.}
\usage{
HAPOST
}
\description{
C-HLI constant \code{HAPOST}
}
\keyword{datasets}
