% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBADVAL}
\alias{HBADVAL}
\title{C-HLI constant \code{HBADVAL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBADVAL
}
\description{
C-HLI constant \code{HBADVAL}
}
\keyword{datasets}
