% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBASRT}
\alias{HBASRT}
\title{C-HLI constant \code{HBASRT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBASRT
}
\description{
C-HLI constant \code{HBASRT}
}
\keyword{datasets}
