% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBATTR}
\alias{HBATTR}
\title{C-HLI constant \code{HBATTR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBATTR
}
\description{
C-HLI constant \code{HBATTR}
}
\keyword{datasets}
