% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBATYP}
\alias{HBATYP}
\title{C-HLI constant \code{HBATYP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBATYP
}
\description{
C-HLI constant \code{HBATYP}
}
\keyword{datasets}
