% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBBASI}
\alias{HBBASI}
\title{C-HLI constant \code{HBBASI}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBBASI
}
\description{
C-HLI constant \code{HBBASI}
}
\keyword{datasets}
