% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBCNTX}
\alias{HBCNTX}
\title{C-HLI constant \code{HBCNTX}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBCNTX
}
\description{
C-HLI constant \code{HBCNTX}
}
\keyword{datasets}
