% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBCONN}
\alias{HBCONN}
\title{C-HLI constant \code{HBCONN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBCONN
}
\description{
C-HLI constant \code{HBCONN}
}
\keyword{datasets}
