% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBDATE}
\alias{HBDATE}
\title{C-HLI constant \code{HBDATE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBDATE
}
\description{
C-HLI constant \code{HBDATE}
}
\keyword{datasets}
