% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBEFOR}
\alias{HBEFOR}
\title{C-HLI constant \code{HBEFOR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBEFOR
}
\description{
C-HLI constant \code{HBEFOR}
}
\keyword{datasets}
