% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBEGIN}
\alias{HBEGIN}
\title{C-HLI constant \code{HBEGIN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBEGIN
}
\description{
C-HLI constant \code{HBEGIN}
}
\keyword{datasets}
