% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBERNG}
\alias{HBERNG}
\title{C-HLI constant \code{HBERNG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBERNG
}
\description{
C-HLI constant \code{HBERNG}
}
\keyword{datasets}
