% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBFILE}
\alias{HBFILE}
\title{C-HLI constant \code{HBFILE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBFILE
}
\description{
C-HLI constant \code{HBFILE}
}
\keyword{datasets}
