% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBFLAB}
\alias{HBFLAB}
\title{C-HLI constant \code{HBFLAB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBFLAB
}
\description{
C-HLI constant \code{HBFLAB}
}
\keyword{datasets}
