% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBFLAG}
\alias{HBFLAG}
\title{C-HLI constant \code{HBFLAG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBFLAG
}
\description{
C-HLI constant \code{HBFLAG}
}
\keyword{datasets}
