% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBFMON}
\alias{HBFMON}
\title{C-HLI constant \code{HBFMON}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBFMON
}
\description{
C-HLI constant \code{HBFMON}
}
\keyword{datasets}
