% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBFREQ}
\alias{HBFREQ}
\title{C-HLI constant \code{HBFREQ}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBFREQ
}
\description{
C-HLI constant \code{HBFREQ}
}
\keyword{datasets}
