% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBFUNC}
\alias{HBFUNC}
\title{C-HLI constant \code{HBFUNC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBFUNC
}
\description{
C-HLI constant \code{HBFUNC}
}
\keyword{datasets}
