% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBGLNM}
\alias{HBGLNM}
\title{C-HLI constant \code{HBGLNM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBGLNM
}
\description{
C-HLI constant \code{HBGLNM}
}
\keyword{datasets}
