% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBGROW}
\alias{HBGROW}
\title{C-HLI constant \code{HBGROW}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBGROW
}
\description{
C-HLI constant \code{HBGROW}
}
\keyword{datasets}
