% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBIMON}
\alias{HBIMON}
\title{C-HLI constant \code{HBIMON}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBIMON
}
\description{
C-HLI constant \code{HBIMON}
}
\keyword{datasets}
