% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBINDX}
\alias{HBINDX}
\title{C-HLI constant \code{HBINDX}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBINDX
}
\description{
C-HLI constant \code{HBINDX}
}
\keyword{datasets}
