% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBMISS}
\alias{HBMISS}
\title{C-HLI constant \code{HBMISS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBMISS
}
\description{
C-HLI constant \code{HBMISS}
}
\keyword{datasets}
