% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBMNOV}
\alias{HBMNOV}
\title{C-HLI constant \code{HBMNOV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBMNOV
}
\description{
C-HLI constant \code{HBMNOV}
}
\keyword{datasets}
