% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBMODE}
\alias{HBMODE}
\title{C-HLI constant \code{HBMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBMODE
}
\description{
C-HLI constant \code{HBMODE}
}
\keyword{datasets}
