% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBMONT}
\alias{HBMONT}
\title{C-HLI constant \code{HBMONT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBMONT
}
\description{
C-HLI constant \code{HBMONT}
}
\keyword{datasets}
