% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBNCHR}
\alias{HBNCHR}
\title{C-HLI constant \code{HBNCHR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBNCHR
}
\description{
C-HLI constant \code{HBNCHR}
}
\keyword{datasets}
