% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBNRNG}
\alias{HBNRNG}
\title{C-HLI constant \code{HBNRNG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBNRNG
}
\description{
C-HLI constant \code{HBNRNG}
}
\keyword{datasets}
