% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBOBJT}
\alias{HBOBJT}
\title{C-HLI constant \code{HBOBJT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBOBJT
}
\description{
C-HLI constant \code{HBOBJT}
}
\keyword{datasets}
