% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBOBSV}
\alias{HBOBSV}
\title{C-HLI constant \code{HBOBSV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBOBSV
}
\description{
C-HLI constant \code{HBOBSV}
}
\keyword{datasets}
