% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBOOLN}
\alias{HBOOLN}
\title{C-HLI constant \code{HBOOLN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBOOLN
}
\description{
C-HLI constant \code{HBOOLN}
}
\keyword{datasets}
