% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBOPTV}
\alias{HBOPTV}
\title{C-HLI constant \code{HBOPTV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBOPTV
}
\description{
C-HLI constant \code{HBOPTV}
}
\keyword{datasets}
