% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBPHAS}
\alias{HBPHAS}
\title{C-HLI constant \code{HBPHAS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBPHAS
}
\description{
C-HLI constant \code{HBPHAS}
}
\keyword{datasets}
