% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBPROD}
\alias{HBPROD}
\title{C-HLI constant \code{HBPROD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBPROD
}
\description{
C-HLI constant \code{HBPROD}
}
\keyword{datasets}
