% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBPRSP}
\alias{HBPRSP}
\title{C-HLI constant \code{HBPRSP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBPRSP
}
\description{
C-HLI constant \code{HBPRSP}
}
\keyword{datasets}
