% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBSBUS}
\alias{HBSBUS}
\title{C-HLI constant \code{HBSBUS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBSBUS
}
\description{
C-HLI constant \code{HBSBUS}
}
\keyword{datasets}
