% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBSDAY}
\alias{HBSDAY}
\title{C-HLI constant \code{HBSDAY}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBSDAY
}
\description{
C-HLI constant \code{HBSDAY}
}
\keyword{datasets}
