% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBSRNG}
\alias{HBSRNG}
\title{C-HLI constant \code{HBSRNG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBSRNG
}
\description{
C-HLI constant \code{HBSRNG}
}
\keyword{datasets}
