% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBSRVR}
\alias{HBSRVR}
\title{C-HLI constant \code{HBSRVR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBSRVR
}
\description{
C-HLI constant \code{HBSRVR}
}
\keyword{datasets}
