% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBSUND}
\alias{HBSUND}
\title{C-HLI constant \code{HBSUND}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBSUND
}
\description{
C-HLI constant \code{HBSUND}
}
\keyword{datasets}
