% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBTIME}
\alias{HBTIME}
\title{C-HLI constant \code{HBTIME}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBTIME
}
\description{
C-HLI constant \code{HBTIME}
}
\keyword{datasets}
