% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBTYPE}
\alias{HBTYPE}
\title{C-HLI constant \code{HBTYPE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBTYPE
}
\description{
C-HLI constant \code{HBTYPE}
}
\keyword{datasets}
