% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBUNIT}
\alias{HBUNIT}
\title{C-HLI constant \code{HBUNIT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBUNIT
}
\description{
C-HLI constant \code{HBUNIT}
}
\keyword{datasets}
