% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBUSER}
\alias{HBUSER}
\title{C-HLI constant \code{HBUSER}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBUSER
}
\description{
C-HLI constant \code{HBUSER}
}
\keyword{datasets}
