% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBUSNS}
\alias{HBUSNS}
\title{C-HLI constant \code{HBUSNS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBUSNS
}
\description{
C-HLI constant \code{HBUSNS}
}
\keyword{datasets}
