% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBV3DATE}
\alias{HBV3DATE}
\title{C-HLI constant \code{HBV3DATE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBV3DATE
}
\description{
C-HLI constant \code{HBV3DATE}
}
\keyword{datasets}
