% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBV3RANGE}
\alias{HBV3RANGE}
\title{C-HLI constant \code{HBV3RANGE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBV3RANGE
}
\description{
C-HLI constant \code{HBV3RANGE}
}
\keyword{datasets}
