% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBV3TYPE}
\alias{HBV3TYPE}
\title{C-HLI constant \code{HBV3TYPE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBV3TYPE
}
\description{
C-HLI constant \code{HBV3TYPE}
}
\keyword{datasets}
