% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HBYEAR}
\alias{HBYEAR}
\title{C-HLI constant \code{HBYEAR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HBYEAR
}
\description{
C-HLI constant \code{HBYEAR}
}
\keyword{datasets}
