% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HCASEX}
\alias{HCASEX}
\title{C-HLI constant \code{HCASEX}}
\format{An object of class \code{integer} of length 1.}
\usage{
HCASEX
}
\description{
C-HLI constant \code{HCASEX}
}
\keyword{datasets}
