% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HCHANL}
\alias{HCHANL}
\title{C-HLI constant \code{HCHANL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HCHANL
}
\description{
C-HLI constant \code{HCHANL}
}
\keyword{datasets}
