% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HCHGAC}
\alias{HCHGAC}
\title{C-HLI constant \code{HCHGAC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HCHGAC
}
\description{
C-HLI constant \code{HCHGAC}
}
\keyword{datasets}
