% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HCLCHN}
\alias{HCLCHN}
\title{C-HLI constant \code{HCLCHN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HCLCHN
}
\description{
C-HLI constant \code{HCLCHN}
}
\keyword{datasets}
