% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HCLNLM}
\alias{HCLNLM}
\title{C-HLI constant \code{HCLNLM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HCLNLM
}
\description{
C-HLI constant \code{HCLNLM}
}
\keyword{datasets}
