% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HCMODE}
\alias{HCMODE}
\title{C-HLI constant \code{HCMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HCMODE
}
\description{
C-HLI constant \code{HCMODE}
}
\keyword{datasets}
