% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HCVTDB}
\alias{HCVTDB}
\title{C-HLI constant \code{HCVTDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HCVTDB
}
\description{
C-HLI constant \code{HCVTDB}
}
\keyword{datasets}
