% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HDAILY}
\alias{HDAILY}
\title{C-HLI constant \code{HDAILY}}
\format{An object of class \code{integer} of length 1.}
\usage{
HDAILY
}
\description{
C-HLI constant \code{HDAILY}
}
\keyword{datasets}
