% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HDBCLM}
\alias{HDBCLM}
\title{C-HLI constant \code{HDBCLM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HDBCLM
}
\description{
C-HLI constant \code{HDBCLM}
}
\keyword{datasets}
