% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HDHOST}
\alias{HDHOST}
\title{C-HLI constant \code{HDHOST}}
\format{An object of class \code{integer} of length 1.}
\usage{
HDHOST
}
\description{
C-HLI constant \code{HDHOST}
}
\keyword{datasets}
