% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HDMODE}
\alias{HDMODE}
\title{C-HLI constant \code{HDMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HDMODE
}
\description{
C-HLI constant \code{HDMODE}
}
\keyword{datasets}
