% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HDPRMC}
\alias{HDPRMC}
\title{C-HLI constant \code{HDPRMC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HDPRMC
}
\description{
C-HLI constant \code{HDPRMC}
}
\keyword{datasets}
