% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HDUTAR}
\alias{HDUTAR}
\title{C-HLI constant \code{HDUTAR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HDUTAR
}
\description{
C-HLI constant \code{HDUTAR}
}
\keyword{datasets}
