% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HEXPIR}
\alias{HEXPIR}
\title{C-HLI constant \code{HEXPIR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HEXPIR
}
\description{
C-HLI constant \code{HEXPIR}
}
\keyword{datasets}
