% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFAMER}
\alias{HFAMER}
\title{C-HLI constant \code{HFAMER}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFAMER
}
\description{
C-HLI constant \code{HFAMER}
}
\keyword{datasets}
