% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFMENV}
\alias{HFMENV}
\title{C-HLI constant \code{HFMENV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFMENV
}
\description{
C-HLI constant \code{HFMENV}
}
\keyword{datasets}
