% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFRMLA}
\alias{HFRMLA}
\title{C-HLI constant \code{HFRMLA}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFRMLA
}
\description{
C-HLI constant \code{HFRMLA}
}
\keyword{datasets}
