% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFRMSYNTAXERR}
\alias{HFRMSYNTAXERR}
\title{C-HLI constant \code{HFRMSYNTAXERR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFRMSYNTAXERR
}
\description{
C-HLI constant \code{HFRMSYNTAXERR}
}
\keyword{datasets}
