% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFTOUT}
\alias{HFTOUT}
\title{C-HLI constant \code{HFTOUT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFTOUT
}
\description{
C-HLI constant \code{HFTOUT}
}
\keyword{datasets}
