% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFUTURE_TYPE}
\alias{HFUTURE_TYPE}
\title{C-HLI constant \code{HFUTURE_TYPE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFUTURE_TYPE
}
\description{
C-HLI constant \code{HFUTURE_TYPE}
}
\keyword{datasets}
