% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYAUG}
\alias{HFYAUG}
\title{C-HLI constant \code{HFYAUG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYAUG
}
\description{
C-HLI constant \code{HFYAUG}
}
\keyword{datasets}
