% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYAUT}
\alias{HFYAUT}
\title{C-HLI constant \code{HFYAUT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYAUT
}
\description{
C-HLI constant \code{HFYAUT}
}
\keyword{datasets}
