% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYDEC}
\alias{HFYDEC}
\title{C-HLI constant \code{HFYDEC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYDEC
}
\description{
C-HLI constant \code{HFYDEC}
}
\keyword{datasets}
