% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYFEB}
\alias{HFYFEB}
\title{C-HLI constant \code{HFYFEB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYFEB
}
\description{
C-HLI constant \code{HFYFEB}
}
\keyword{datasets}
