% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYFST}
\alias{HFYFST}
\title{C-HLI constant \code{HFYFST}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYFST
}
\description{
C-HLI constant \code{HFYFST}
}
\keyword{datasets}
