% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYJAN}
\alias{HFYJAN}
\title{C-HLI constant \code{HFYJAN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYJAN
}
\description{
C-HLI constant \code{HFYJAN}
}
\keyword{datasets}
