% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYJUL}
\alias{HFYJUL}
\title{C-HLI constant \code{HFYJUL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYJUL
}
\description{
C-HLI constant \code{HFYJUL}
}
\keyword{datasets}
