% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYJUN}
\alias{HFYJUN}
\title{C-HLI constant \code{HFYJUN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYJUN
}
\description{
C-HLI constant \code{HFYJUN}
}
\keyword{datasets}
