% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYLST}
\alias{HFYLST}
\title{C-HLI constant \code{HFYLST}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYLST
}
\description{
C-HLI constant \code{HFYLST}
}
\keyword{datasets}
