% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYMAR}
\alias{HFYMAR}
\title{C-HLI constant \code{HFYMAR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYMAR
}
\description{
C-HLI constant \code{HFYMAR}
}
\keyword{datasets}
