% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYMAY}
\alias{HFYMAY}
\title{C-HLI constant \code{HFYMAY}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYMAY
}
\description{
C-HLI constant \code{HFYMAY}
}
\keyword{datasets}
