% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYNOV}
\alias{HFYNOV}
\title{C-HLI constant \code{HFYNOV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYNOV
}
\description{
C-HLI constant \code{HFYNOV}
}
\keyword{datasets}
