% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYOCT}
\alias{HFYOCT}
\title{C-HLI constant \code{HFYOCT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYOCT
}
\description{
C-HLI constant \code{HFYOCT}
}
\keyword{datasets}
