% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HFYSEP}
\alias{HFYSEP}
\title{C-HLI constant \code{HFYSEP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HFYSEP
}
\description{
C-HLI constant \code{HFYSEP}
}
\keyword{datasets}
