% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HGLFOR}
\alias{HGLFOR}
\title{C-HLI constant \code{HGLFOR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HGLFOR
}
\description{
C-HLI constant \code{HGLFOR}
}
\keyword{datasets}
