% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HGLNAM}
\alias{HGLNAM}
\title{C-HLI constant \code{HGLNAM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HGLNAM
}
\description{
C-HLI constant \code{HGLNAM}
}
\keyword{datasets}
