% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HIFAIL}
\alias{HIFAIL}
\title{C-HLI constant \code{HIFAIL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HIFAIL
}
\description{
C-HLI constant \code{HIFAIL}
}
\keyword{datasets}
