% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HINITD}
\alias{HINITD}
\title{C-HLI constant \code{HINITD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HINITD
}
\description{
C-HLI constant \code{HINITD}
}
\keyword{datasets}
