% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HINTVL}
\alias{HINTVL}
\title{C-HLI constant \code{HINTVL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HINTVL
}
\description{
C-HLI constant \code{HINTVL}
}
\keyword{datasets}
