% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HINUSE}
\alias{HINUSE}
\title{C-HLI constant \code{HINUSE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HINUSE
}
\description{
C-HLI constant \code{HINUSE}
}
\keyword{datasets}
