% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLI_MAX_FAME_CMD}
\alias{HLI_MAX_FAME_CMD}
\title{C-HLI constant \code{HLI_MAX_FAME_CMD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLI_MAX_FAME_CMD
}
\description{
C-HLI constant \code{HLI_MAX_FAME_CMD}
}
\keyword{datasets}
