% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLI_MAX_FAME_INPUT_SIZE}
\alias{HLI_MAX_FAME_INPUT_SIZE}
\title{C-HLI constant \code{HLI_MAX_FAME_INPUT_SIZE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLI_MAX_FAME_INPUT_SIZE
}
\description{
C-HLI constant \code{HLI_MAX_FAME_INPUT_SIZE}
}
\keyword{datasets}
