% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLI_MAX_STR_LEN}
\alias{HLI_MAX_STR_LEN}
\title{C-HLI constant \code{HLI_MAX_STR_LEN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLI_MAX_STR_LEN
}
\description{
C-HLI constant \code{HLI_MAX_STR_LEN}
}
\keyword{datasets}
