% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLICFL}
\alias{HLICFL}
\title{C-HLI constant \code{HLICFL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLICFL
}
\description{
C-HLI constant \code{HLICFL}
}
\keyword{datasets}
