% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLICNS}
\alias{HLICNS}
\title{C-HLI constant \code{HLICNS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLICNS
}
\description{
C-HLI constant \code{HLICNS}
}
\keyword{datasets}
