% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLOCKD}
\alias{HLOCKD}
\title{C-HLI constant \code{HLOCKD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLOCKD
}
\description{
C-HLI constant \code{HLOCKD}
}
\keyword{datasets}
