% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLRESV}
\alias{HLRESV}
\title{C-HLI constant \code{HLRESV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLRESV
}
\description{
C-HLI constant \code{HLRESV}
}
\keyword{datasets}
