% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HLSERV}
\alias{HLSERV}
\title{C-HLI constant \code{HLSERV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HLSERV
}
\description{
C-HLI constant \code{HLSERV}
}
\keyword{datasets}
