% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HMAXDB}
\alias{HMAXDB}
\title{C-HLI constant \code{HMAXDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HMAXDB
}
\description{
C-HLI constant \code{HMAXDB}
}
\keyword{datasets}
