% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HMAXSCMD}
\alias{HMAXSCMD}
\title{C-HLI constant \code{HMAXSCMD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HMAXSCMD
}
\description{
C-HLI constant \code{HMAXSCMD}
}
\keyword{datasets}
