% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HMFILE}
\alias{HMFILE}
\title{C-HLI constant \code{HMFILE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HMFILE
}
\description{
C-HLI constant \code{HMFILE}
}
\keyword{datasets}
