% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HMGVAL}
\alias{HMGVAL}
\title{C-HLI constant \code{HMGVAL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HMGVAL
}
\description{
C-HLI constant \code{HMGVAL}
}
\keyword{datasets}
