% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HMONTH}
\alias{HMONTH}
\title{C-HLI constant \code{HMONTH}}
\format{An object of class \code{integer} of length 1.}
\usage{
HMONTH
}
\description{
C-HLI constant \code{HMONTH}
}
\keyword{datasets}
