% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HMPOST}
\alias{HMPOST}
\title{C-HLI constant \code{HMPOST}}
\format{An object of class \code{integer} of length 1.}
\usage{
HMPOST
}
\description{
C-HLI constant \code{HMPOST}
}
\keyword{datasets}
