% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNAMEL}
\alias{HNAMEL}
\title{C-HLI constant \code{HNAMEL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNAMEL
}
\description{
C-HLI constant \code{HNAMEL}
}
\keyword{datasets}
