% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNAMLEN}
\alias{HNAMLEN}
\title{C-HLI constant \code{HNAMLEN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNAMLEN
}
\description{
C-HLI constant \code{HNAMLEN}
}
\keyword{datasets}
