% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNAMLEN_V4}
\alias{HNAMLEN_V4}
\title{C-HLI constant \code{HNAMLEN_V4}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNAMLEN_V4
}
\description{
C-HLI constant \code{HNAMLEN_V4}
}
\keyword{datasets}
