% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNAMSIZ}
\alias{HNAMSIZ}
\title{C-HLI constant \code{HNAMSIZ}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNAMSIZ
}
\description{
C-HLI constant \code{HNAMSIZ}
}
\keyword{datasets}
