% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNAVAL}
\alias{HNAVAL}
\title{C-HLI constant \code{HNAVAL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNAVAL
}
\description{
C-HLI constant \code{HNAVAL}
}
\keyword{datasets}
