% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNBACK}
\alias{HNBACK}
\title{C-HLI constant \code{HNBACK}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNBACK
}
\description{
C-HLI constant \code{HNBACK}
}
\keyword{datasets}
