% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNCONN}
\alias{HNCONN}
\title{C-HLI constant \code{HNCONN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNCONN
}
\description{
C-HLI constant \code{HNCONN}
}
\keyword{datasets}
