% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNCVAL}
\alias{HNCVAL}
\title{C-HLI constant \code{HNCVAL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNCVAL
}
\description{
C-HLI constant \code{HNCVAL}
}
\keyword{datasets}
