% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNDVAL}
\alias{HNDVAL}
\title{C-HLI constant \code{HNDVAL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNDVAL
}
\description{
C-HLI constant \code{HNDVAL}
}
\keyword{datasets}
