% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNEMPT}
\alias{HNEMPT}
\title{C-HLI constant \code{HNEMPT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNEMPT
}
\description{
C-HLI constant \code{HNEMPT}
}
\keyword{datasets}
