% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNETCN}
\alias{HNETCN}
\title{C-HLI constant \code{HNETCN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNETCN
}
\description{
C-HLI constant \code{HNETCN}
}
\keyword{datasets}
