% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNFAME}
\alias{HNFAME}
\title{C-HLI constant \code{HNFAME}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNFAME
}
\description{
C-HLI constant \code{HNFAME}
}
\keyword{datasets}
