% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNFILE}
\alias{HNFILE}
\title{C-HLI constant \code{HNFILE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNFILE
}
\description{
C-HLI constant \code{HNFILE}
}
\keyword{datasets}
