% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNFMDB}
\alias{HNFMDB}
\title{C-HLI constant \code{HNFMDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNFMDB
}
\description{
C-HLI constant \code{HNFMDB}
}
\keyword{datasets}
