% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNFSRCLONG}
\alias{HNFSRCLONG}
\title{C-HLI constant \code{HNFSRCLONG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNFSRCLONG
}
\description{
C-HLI constant \code{HNFSRCLONG}
}
\keyword{datasets}
