% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNINIT}
\alias{HNINIT}
\title{C-HLI constant \code{HNINIT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNINIT
}
\description{
C-HLI constant \code{HNINIT}
}
\keyword{datasets}
