% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNLALL}
\alias{HNLALL}
\title{C-HLI constant \code{HNLALL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNLALL
}
\description{
C-HLI constant \code{HNLALL}
}
\keyword{datasets}
