% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNLOCL}
\alias{HNLOCL}
\title{C-HLI constant \code{HNLOCL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNLOCL
}
\description{
C-HLI constant \code{HNLOCL}
}
\keyword{datasets}
