% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNMVAL}
\alias{HNMVAL}
\title{C-HLI constant \code{HNMVAL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNMVAL
}
\description{
C-HLI constant \code{HNMVAL}
}
\keyword{datasets}
