% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNOFORMV3}
\alias{HNOFORMV3}
\title{C-HLI constant \code{HNOFORMV3}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNOFORMV3
}
\description{
C-HLI constant \code{HNOFORMV3}
}
\keyword{datasets}
