% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNOMAP}
\alias{HNOMAP}
\title{C-HLI constant \code{HNOMAP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNOMAP
}
\description{
C-HLI constant \code{HNOMAP}
}
\keyword{datasets}
