% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNOMEM}
\alias{HNOMEM}
\title{C-HLI constant \code{HNOMEM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNOMEM
}
\description{
C-HLI constant \code{HNOMEM}
}
\keyword{datasets}
