% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNOOBJ}
\alias{HNOOBJ}
\title{C-HLI constant \code{HNOOBJ}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNOOBJ
}
\description{
C-HLI constant \code{HNOOBJ}
}
\keyword{datasets}
