% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNPOST}
\alias{HNPOST}
\title{C-HLI constant \code{HNPOST}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNPOST
}
\description{
C-HLI constant \code{HNPOST}
}
\keyword{datasets}
