% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNRESW}
\alias{HNRESW}
\title{C-HLI constant \code{HNRESW}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNRESW
}
\description{
C-HLI constant \code{HNRESW}
}
\keyword{datasets}
