% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNSUPDB}
\alias{HNSUPDB}
\title{C-HLI constant \code{HNSUPDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNSUPDB
}
\description{
C-HLI constant \code{HNSUPDB}
}
\keyword{datasets}
