% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNSUPP}
\alias{HNSUPP}
\title{C-HLI constant \code{HNSUPP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNSUPP
}
\description{
C-HLI constant \code{HNSUPP}
}
\keyword{datasets}
