% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNTMIS}
\alias{HNTMIS}
\title{C-HLI constant \code{HNTMIS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNTMIS
}
\description{
C-HLI constant \code{HNTMIS}
}
\keyword{datasets}
