% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNTOOLONG}
\alias{HNTOOLONG}
\title{C-HLI constant \code{HNTOOLONG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNTOOLONG
}
\description{
C-HLI constant \code{HNTOOLONG}
}
\keyword{datasets}
