% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNTSAFE}
\alias{HNTSAFE}
\title{C-HLI constant \code{HNTSAFE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNTSAFE
}
\description{
C-HLI constant \code{HNTSAFE}
}
\keyword{datasets}
