% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNTWIC}
\alias{HNTWIC}
\title{C-HLI constant \code{HNTWIC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNTWIC
}
\description{
C-HLI constant \code{HNTWIC}
}
\keyword{datasets}
