% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNUFRD}
\alias{HNUFRD}
\title{C-HLI constant \code{HNUFRD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNUFRD
}
\description{
C-HLI constant \code{HNUFRD}
}
\keyword{datasets}
