% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNULLP}
\alias{HNULLP}
\title{C-HLI constant \code{HNULLP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNULLP
}
\description{
C-HLI constant \code{HNULLP}
}
\keyword{datasets}
