% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNUMRC}
\alias{HNUMRC}
\title{C-HLI constant \code{HNUMRC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNUMRC
}
\description{
C-HLI constant \code{HNUMRC}
}
\keyword{datasets}
