% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNWFEA}
\alias{HNWFEA}
\title{C-HLI constant \code{HNWFEA}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNWFEA
}
\description{
C-HLI constant \code{HNWFEA}
}
\keyword{datasets}
