% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HNWILD}
\alias{HNWILD}
\title{C-HLI constant \code{HNWILD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HNWILD
}
\description{
C-HLI constant \code{HNWILD}
}
\keyword{datasets}
