% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBANN}
\alias{HOBANN}
\title{C-HLI constant \code{HOBANN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBANN
}
\description{
C-HLI constant \code{HOBANN}
}
\keyword{datasets}
