% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBAVG}
\alias{HOBAVG}
\title{C-HLI constant \code{HOBAVG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBAVG
}
\description{
C-HLI constant \code{HOBAVG}
}
\keyword{datasets}
