% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBBEG}
\alias{HOBBEG}
\title{C-HLI constant \code{HOBBEG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBBEG
}
\description{
C-HLI constant \code{HOBBEG}
}
\keyword{datasets}
