% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBEND}
\alias{HOBEND}
\title{C-HLI constant \code{HOBEND}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBEND
}
\description{
C-HLI constant \code{HOBEND}
}
\keyword{datasets}
