% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBFRM}
\alias{HOBFRM}
\title{C-HLI constant \code{HOBFRM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBFRM
}
\description{
C-HLI constant \code{HOBFRM}
}
\keyword{datasets}
