% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBJSIZE}
\alias{HOBJSIZE}
\title{C-HLI constant \code{HOBJSIZE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBJSIZE
}
\description{
C-HLI constant \code{HOBJSIZE}
}
\keyword{datasets}
