% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBSFUNC}
\alias{HOBSFUNC}
\title{C-HLI constant \code{HOBSFUNC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBSFUNC
}
\description{
C-HLI constant \code{HOBSFUNC}
}
\keyword{datasets}
