% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBSUM}
\alias{HOBSUM}
\title{C-HLI constant \code{HOBSUM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBSUM
}
\description{
C-HLI constant \code{HOBSUM}
}
\keyword{datasets}
