% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOBUND}
\alias{HOBUND}
\title{C-HLI constant \code{HOBUND}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOBUND
}
\description{
C-HLI constant \code{HOBUND}
}
\keyword{datasets}
