% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOLDDB}
\alias{HOLDDB}
\title{C-HLI constant \code{HOLDDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOLDDB
}
\description{
C-HLI constant \code{HOLDDB}
}
\keyword{datasets}
