% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOMODE}
\alias{HOMODE}
\title{C-HLI constant \code{HOMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOMODE
}
\description{
C-HLI constant \code{HOMODE}
}
\keyword{datasets}
