% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOPEND}
\alias{HOPEND}
\title{C-HLI constant \code{HOPEND}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOPEND
}
\description{
C-HLI constant \code{HOPEND}
}
\keyword{datasets}
