% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HOPENW}
\alias{HOPENW}
\title{C-HLI constant \code{HOPENW}}
\format{An object of class \code{integer} of length 1.}
\usage{
HOPENW
}
\description{
C-HLI constant \code{HOPENW}
}
\keyword{datasets}
