% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HP1REQ}
\alias{HP1REQ}
\title{C-HLI constant \code{HP1REQ}}
\format{An object of class \code{integer} of length 1.}
\usage{
HP1REQ
}
\description{
C-HLI constant \code{HP1REQ}
}
\keyword{datasets}
