% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HP2REQ}
\alias{HP2REQ}
\title{C-HLI constant \code{HP2REQ}}
\format{An object of class \code{integer} of length 1.}
\usage{
HP2REQ
}
\description{
C-HLI constant \code{HP2REQ}
}
\keyword{datasets}
