% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HPRECN}
\alias{HPRECN}
\title{C-HLI constant \code{HPRECN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HPRECN
}
\description{
C-HLI constant \code{HPRECN}
}
\keyword{datasets}
