% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HPWWOU}
\alias{HPWWOU}
\title{C-HLI constant \code{HPWWOU}}
\format{An object of class \code{integer} of length 1.}
\usage{
HPWWOU
}
\description{
C-HLI constant \code{HPWWOU}
}
\keyword{datasets}
