% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HQTDEC}
\alias{HQTDEC}
\title{C-HLI constant \code{HQTDEC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HQTDEC
}
\description{
C-HLI constant \code{HQTDEC}
}
\keyword{datasets}
