% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HQTNOV}
\alias{HQTNOV}
\title{C-HLI constant \code{HQTNOV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HQTNOV
}
\description{
C-HLI constant \code{HQTNOV}
}
\keyword{datasets}
