% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HQTOCT}
\alias{HQTOCT}
\title{C-HLI constant \code{HQTOCT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HQTOCT
}
\description{
C-HLI constant \code{HQTOCT}
}
\keyword{datasets}
