% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HQUOTA}
\alias{HQUOTA}
\title{C-HLI constant \code{HQUOTA}}
\format{An object of class \code{integer} of length 1.}
\usage{
HQUOTA
}
\description{
C-HLI constant \code{HQUOTA}
}
\keyword{datasets}
