% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRDISTRANS}
\alias{HRDISTRANS}
\title{C-HLI constant \code{HRDISTRANS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRDISTRANS
}
\description{
C-HLI constant \code{HRDISTRANS}
}
\keyword{datasets}
