% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HREADO}
\alias{HREADO}
\title{C-HLI constant \code{HREADO}}
\format{An object of class \code{integer} of length 1.}
\usage{
HREADO
}
\description{
C-HLI constant \code{HREADO}
}
\keyword{datasets}
