% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HREMSUP}
\alias{HREMSUP}
\title{C-HLI constant \code{HREMSUP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HREMSUP
}
\description{
C-HLI constant \code{HREMSUP}
}
\keyword{datasets}
