% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRESFD}
\alias{HRESFD}
\title{C-HLI constant \code{HRESFD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRESFD
}
\description{
C-HLI constant \code{HRESFD}
}
\keyword{datasets}
