% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRHANDINV}
\alias{HRHANDINV}
\title{C-HLI constant \code{HRHANDINV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRHANDINV
}
\description{
C-HLI constant \code{HRHANDINV}
}
\keyword{datasets}
