% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRKEYINV}
\alias{HRKEYINV}
\title{C-HLI constant \code{HRKEYINV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRKEYINV
}
\description{
C-HLI constant \code{HRKEYINV}
}
\keyword{datasets}
