% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRMKEY}
\alias{HRMKEY}
\title{C-HLI constant \code{HRMKEY}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRMKEY
}
\description{
C-HLI constant \code{HRMKEY}
}
\keyword{datasets}
