% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRMODE}
\alias{HRMODE}
\title{C-HLI constant \code{HRMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRMODE
}
\description{
C-HLI constant \code{HRMODE}
}
\keyword{datasets}
