% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRMTDB}
\alias{HRMTDB}
\title{C-HLI constant \code{HRMTDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRMTDB
}
\description{
C-HLI constant \code{HRMTDB}
}
\keyword{datasets}
