% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRNEXI}
\alias{HRNEXI}
\title{C-HLI constant \code{HRNEXI}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRNEXI
}
\description{
C-HLI constant \code{HRNEXI}
}
\keyword{datasets}
