% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HROSCONN}
\alias{HROSCONN}
\title{C-HLI constant \code{HROSCONN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HROSCONN
}
\description{
C-HLI constant \code{HROSCONN}
}
\keyword{datasets}
