% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HROSTRANS}
\alias{HROSTRANS}
\title{C-HLI constant \code{HROSTRANS}}
\format{An object of class \code{integer} of length 1.}
\usage{
HROSTRANS
}
\description{
C-HLI constant \code{HROSTRANS}
}
\keyword{datasets}
