% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HRUSERINV}
\alias{HRUSERINV}
\title{C-HLI constant \code{HRUSERINV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HRUSERINV
}
\description{
C-HLI constant \code{HRUSERINV}
}
\keyword{datasets}
