% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSCALA}
\alias{HSCALA}
\title{C-HLI constant \code{HSCALA}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSCALA
}
\description{
C-HLI constant \code{HSCALA}
}
\keyword{datasets}
