% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSCLLM}
\alias{HSCLLM}
\title{C-HLI constant \code{HSCLLM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSCLLM
}
\description{
C-HLI constant \code{HSCLLM}
}
\keyword{datasets}
