% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSERIE}
\alias{HSERIE}
\title{C-HLI constant \code{HSERIE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSERIE
}
\description{
C-HLI constant \code{HSERIE}
}
\keyword{datasets}
