% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSERVR}
\alias{HSERVR}
\title{C-HLI constant \code{HSERVR}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSERVR
}
\description{
C-HLI constant \code{HSERVR}
}
\keyword{datasets}
