% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMAUG}
\alias{HSMAUG}
\title{C-HLI constant \code{HSMAUG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMAUG
}
\description{
C-HLI constant \code{HSMAUG}
}
\keyword{datasets}
