% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMDEC}
\alias{HSMDEC}
\title{C-HLI constant \code{HSMDEC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMDEC
}
\description{
C-HLI constant \code{HSMDEC}
}
\keyword{datasets}
