% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMFIL}
\alias{HSMFIL}
\title{C-HLI constant \code{HSMFIL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMFIL
}
\description{
C-HLI constant \code{HSMFIL}
}
\keyword{datasets}
