% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMJUL}
\alias{HSMJUL}
\title{C-HLI constant \code{HSMJUL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMJUL
}
\description{
C-HLI constant \code{HSMJUL}
}
\keyword{datasets}
