% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMNOV}
\alias{HSMNOV}
\title{C-HLI constant \code{HSMNOV}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMNOV
}
\description{
C-HLI constant \code{HSMNOV}
}
\keyword{datasets}
