% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMOCT}
\alias{HSMOCT}
\title{C-HLI constant \code{HSMOCT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMOCT
}
\description{
C-HLI constant \code{HSMOCT}
}
\keyword{datasets}
