% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMODE}
\alias{HSMODE}
\title{C-HLI constant \code{HSMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMODE
}
\description{
C-HLI constant \code{HSMODE}
}
\keyword{datasets}
