% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSMSEP}
\alias{HSMSEP}
\title{C-HLI constant \code{HSMSEP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSMSEP
}
\description{
C-HLI constant \code{HSMSEP}
}
\keyword{datasets}
