% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSNFIL}
\alias{HSNFIL}
\title{C-HLI constant \code{HSNFIL}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSNFIL
}
\description{
C-HLI constant \code{HSNFIL}
}
\keyword{datasets}
