% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSPCDB}
\alias{HSPCDB}
\title{C-HLI constant \code{HSPCDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSPCDB
}
\description{
C-HLI constant \code{HSPCDB}
}
\keyword{datasets}
