% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSRVST}
\alias{HSRVST}
\title{C-HLI constant \code{HSRVST}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSRVST
}
\description{
C-HLI constant \code{HSRVST}
}
\keyword{datasets}
