% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSTRNG}
\alias{HSTRNG}
\title{C-HLI constant \code{HSTRNG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSTRNG
}
\description{
C-HLI constant \code{HSTRNG}
}
\keyword{datasets}
