% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSUCC}
\alias{HSUCC}
\title{C-HLI constant \code{HSUCC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSUCC
}
\description{
C-HLI constant \code{HSUCC}
}
\examples{
status <- Integer(-1L)
cfmini(status)
if(status$value!=HSUCC){
  cat(sprintf("cfmini(status=\%d)",status$value))
}

}
\keyword{datasets}
