% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HSUSPN}
\alias{HSUSPN}
\title{C-HLI constant \code{HSUSPN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HSUSPN
}
\description{
C-HLI constant \code{HSUSPN}
}
\keyword{datasets}
