% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HTMOUT}
\alias{HTMOUT}
\title{C-HLI constant \code{HTMOUT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HTMOUT
}
\description{
C-HLI constant \code{HTMOUT}
}
\keyword{datasets}
