% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HTRUNC}
\alias{HTRUNC}
\title{C-HLI constant \code{HTRUNC}}
\format{An object of class \code{integer} of length 1.}
\usage{
HTRUNC
}
\description{
C-HLI constant \code{HTRUNC}
}
\keyword{datasets}
