% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HTRUNCDB}
\alias{HTRUNCDB}
\title{C-HLI constant \code{HTRUNCDB}}
\format{An object of class \code{integer} of length 1.}
\usage{
HTRUNCDB
}
\description{
C-HLI constant \code{HTRUNCDB}
}
\keyword{datasets}
