% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HTWICM}
\alias{HTWICM}
\title{C-HLI constant \code{HTWICM}}
\format{An object of class \code{integer} of length 1.}
\usage{
HTWICM
}
\description{
C-HLI constant \code{HTWICM}
}
\keyword{datasets}
