% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUMODE}
\alias{HUMODE}
\title{C-HLI constant \code{HUMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUMODE
}
\description{
C-HLI constant \code{HUMODE}
}
\keyword{datasets}
