% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUNCHG}
\alias{HUNCHG}
\title{C-HLI constant \code{HUNCHG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUNCHG
}
\description{
C-HLI constant \code{HUNCHG}
}
\keyword{datasets}
