% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUNDFT}
\alias{HUNDFT}
\title{C-HLI constant \code{HUNDFT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUNDFT
}
\description{
C-HLI constant \code{HUNDFT}
}
\keyword{datasets}
