% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUNDFX}
\alias{HUNDFX}
\title{C-HLI constant \code{HUNDFX}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUNDFX
}
\description{
C-HLI constant \code{HUNDFX}
}
\keyword{datasets}
