% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUNEXP}
\alias{HUNEXP}
\title{C-HLI constant \code{HUNEXP}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUNEXP
}
\description{
C-HLI constant \code{HUNEXP}
}
\keyword{datasets}
