% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUNSUPPORTED_FAME_VER}
\alias{HUNSUPPORTED_FAME_VER}
\title{C-HLI constant \code{HUNSUPPORTED_FAME_VER}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUNSUPPORTED_FAME_VER
}
\description{
C-HLI constant \code{HUNSUPPORTED_FAME_VER}
}
\keyword{datasets}
