% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUPDRD}
\alias{HUPDRD}
\title{C-HLI constant \code{HUPDRD}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUPDRD
}
\description{
C-HLI constant \code{HUPDRD}
}
\keyword{datasets}
