% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HUSRPWTOOLONG}
\alias{HUSRPWTOOLONG}
\title{C-HLI constant \code{HUSRPWTOOLONG}}
\format{An object of class \code{integer} of length 1.}
\usage{
HUSRPWTOOLONG
}
\description{
C-HLI constant \code{HUSRPWTOOLONG}
}
\keyword{datasets}
