% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWAFRI}
\alias{HWAFRI}
\title{C-HLI constant \code{HWAFRI}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWAFRI
}
\description{
C-HLI constant \code{HWAFRI}
}
\keyword{datasets}
