% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWAMON}
\alias{HWAMON}
\title{C-HLI constant \code{HWAMON}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWAMON
}
\description{
C-HLI constant \code{HWAMON}
}
\keyword{datasets}
