% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWASAT}
\alias{HWASAT}
\title{C-HLI constant \code{HWASAT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWASAT
}
\description{
C-HLI constant \code{HWASAT}
}
\keyword{datasets}
