% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWATHU}
\alias{HWATHU}
\title{C-HLI constant \code{HWATHU}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWATHU
}
\description{
C-HLI constant \code{HWATHU}
}
\keyword{datasets}
