% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWATUE}
\alias{HWATUE}
\title{C-HLI constant \code{HWATUE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWATUE
}
\description{
C-HLI constant \code{HWATUE}
}
\keyword{datasets}
