% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWAWED}
\alias{HWAWED}
\title{C-HLI constant \code{HWAWED}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWAWED
}
\description{
C-HLI constant \code{HWAWED}
}
\keyword{datasets}
