% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWBFRI}
\alias{HWBFRI}
\title{C-HLI constant \code{HWBFRI}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWBFRI
}
\description{
C-HLI constant \code{HWBFRI}
}
\keyword{datasets}
