% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWBMON}
\alias{HWBMON}
\title{C-HLI constant \code{HWBMON}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWBMON
}
\description{
C-HLI constant \code{HWBMON}
}
\keyword{datasets}
