% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWBSAT}
\alias{HWBSAT}
\title{C-HLI constant \code{HWBSAT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWBSAT
}
\description{
C-HLI constant \code{HWBSAT}
}
\keyword{datasets}
