% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWBSUN}
\alias{HWBSUN}
\title{C-HLI constant \code{HWBSUN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWBSUN
}
\description{
C-HLI constant \code{HWBSUN}
}
\keyword{datasets}
