% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWBTHU}
\alias{HWBTHU}
\title{C-HLI constant \code{HWBTHU}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWBTHU
}
\description{
C-HLI constant \code{HWBTHU}
}
\keyword{datasets}
