% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWBTUE}
\alias{HWBTUE}
\title{C-HLI constant \code{HWBTUE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWBTUE
}
\description{
C-HLI constant \code{HWBTUE}
}
\keyword{datasets}
