% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWBWED}
\alias{HWBWED}
\title{C-HLI constant \code{HWBWED}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWBWED
}
\description{
C-HLI constant \code{HWBWED}
}
\keyword{datasets}
