% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWEEK_PATTERN}
\alias{HWEEK_PATTERN}
\title{C-HLI constant \code{HWEEK_PATTERN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWEEK_PATTERN
}
\description{
C-HLI constant \code{HWEEK_PATTERN}
}
\keyword{datasets}
