% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKFRI}
\alias{HWKFRI}
\title{C-HLI constant \code{HWKFRI}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKFRI
}
\description{
C-HLI constant \code{HWKFRI}
}
\keyword{datasets}
