% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKMON}
\alias{HWKMON}
\title{C-HLI constant \code{HWKMON}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKMON
}
\description{
C-HLI constant \code{HWKMON}
}
\keyword{datasets}
