% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKOPN}
\alias{HWKOPN}
\title{C-HLI constant \code{HWKOPN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKOPN
}
\description{
C-HLI constant \code{HWKOPN}
}
\keyword{datasets}
