% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKSAT}
\alias{HWKSAT}
\title{C-HLI constant \code{HWKSAT}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKSAT
}
\description{
C-HLI constant \code{HWKSAT}
}
\keyword{datasets}
