% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKSUN}
\alias{HWKSUN}
\title{C-HLI constant \code{HWKSUN}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKSUN
}
\description{
C-HLI constant \code{HWKSUN}
}
\keyword{datasets}
