% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKTHU}
\alias{HWKTHU}
\title{C-HLI constant \code{HWKTHU}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKTHU
}
\description{
C-HLI constant \code{HWKTHU}
}
\keyword{datasets}
