% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKTUE}
\alias{HWKTUE}
\title{C-HLI constant \code{HWKTUE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKTUE
}
\description{
C-HLI constant \code{HWKTUE}
}
\keyword{datasets}
