% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWKWED}
\alias{HWKWED}
\title{C-HLI constant \code{HWKWED}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWKWED
}
\description{
C-HLI constant \code{HWKWED}
}
\keyword{datasets}
