% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWMODE}
\alias{HWMODE}
\title{C-HLI constant \code{HWMODE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWMODE
}
\description{
C-HLI constant \code{HWMODE}
}
\keyword{datasets}
