% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{data}
\name{HWRITE}
\alias{HWRITE}
\title{C-HLI constant \code{HWRITE}}
\format{An object of class \code{integer} of length 1.}
\usage{
HWRITE
}
\description{
C-HLI constant \code{HWRITE}
}
\keyword{datasets}
