% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{class}
\name{Integer-class}
\alias{Integer-class}
\alias{Integer}
\title{\code{Integer}}
\description{
A reference class to implement mutable integer data.
}
\section{Fields}{

\describe{
\item{\code{value}}{An integer vector.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(i = 0L)}}{Initialize an integer array.}
}}

\examples{
status <- Integer(-1)

}
