% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{class}
\name{Numeric-class}
\alias{Numeric-class}
\alias{Numeric}
\title{\code{Numeric}}
\description{
A reference class to implement mutable numeric data.
}
\section{Fields}{

\describe{
\item{\code{value}}{A numeric vector.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(x = 0)}}{Initialize a numeric vector.}
}}

\examples{
ver <- Numeric(-1)

}
