% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhli.R
\docType{package}
\name{rhli}
\alias{rhli}
\alias{rhli-package}
\title{rhli: R access to the FAME C-HLI.}
\description{
Complete access from R to the FIS MarketMap
C-Toolkit (FAME C-HLI).
}
\details{
FAME is a fully integrated software and database
management system from FIS that provides the following capabilities:
\itemize{
\item Time series and cross-sectional data management;
\item Financial calculation, data analysis, econometrics, and forecasting;
\item Table generation and detailed multicolor, presentation-quality report writing;
\item Multicolor, presentation-quality graphics;
\item "What-if" analysis;
\item Application development and structured programming;
\item Data transfer to and from other applications;
\item Tools for building customized graphical user interfaces.
}

Programs should:
\itemize{
\item Initialize with \code{\link{cfmini}} first.
\item Display the version number with \code{\link{cfmver}}.
\item \emph{( do other work )}
\item Finish with \code{\link{cfmfin}}.
}

There are three data classes for use with rhli:
\itemize{
\item \code{\link{Character}} 
\item \code{\link{Integer}} 
\item \code{\link{Numeric}}
}
}
